+++ hash/Cargo.toml
@@ -38,10 +38,6 @@ solana-frozen-abi-macro = { workspace = true, optional = true, features = [
 ] }
 solana-sanitize = { workspace = true }
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-js-sys = { workspace = true }
-wasm-bindgen = { workspace = true }
-
 [dev-dependencies]
 bs58 = { workspace = true, default-features = false, features = ["alloc"] }
