+++ hash/src/lib.rs
@@ -3,13 +3,13 @@
 #![cfg_attr(feature = "frozen-abi", feature(min_specialization))]
 #[cfg(feature = "borsh")]
 use borsh::{BorshDeserialize, BorshSchema, BorshSerialize};
-#[cfg(any(feature = "std", target_arch = "wasm32"))]
+#[cfg(feature = "std")]
 extern crate std;
 #[cfg(feature = "bytemuck")]
 use bytemuck_derive::{Pod, Zeroable};
 #[cfg(feature = "serde")]
 use serde_derive::{Deserialize, Serialize};
-#[cfg(any(all(feature = "borsh", feature = "std"), target_arch = "wasm32"))]
+#[cfg(all(feature = "borsh", feature = "std"))]
 use std::string::ToString;
 use {
     core::{
@@ -17,12 +17,6 @@ use {
         str::{from_utf8_unchecked, FromStr},
     },
     solana_sanitize::Sanitize,
-};
-#[cfg(target_arch = "wasm32")]
-use {
-    js_sys::{Array, Uint8Array},
-    std::{boxed::Box, format, string::String, vec},
-    wasm_bindgen::{prelude::*, JsCast},
 };
 
 /// Size of a hash in bytes.
@@ -37,7 +31,6 @@ pub const MAX_BASE58_LEN: usize = 44;
 ///
 /// [SHA-256]: https://en.wikipedia.org/wiki/SHA-2
 /// [blake3]: https://github.com/BLAKE3-team/BLAKE3
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 #[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
 #[cfg_attr(
     feature = "borsh",
@@ -145,63 +138,6 @@ impl Hash {
 
     pub const fn as_bytes(&self) -> &[u8; HASH_BYTES] {
         &self.0
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-#[allow(non_snake_case)]
-#[wasm_bindgen]
-impl Hash {
-    /// Create a new Hash object
-    ///
-    /// * `value` - optional hash as a base58 encoded string, `Uint8Array`, `[number]`
-    #[wasm_bindgen(constructor)]
-    pub fn constructor(value: JsValue) -> Result<Hash, JsValue> {
-        if let Some(base58_str) = value.as_string() {
-            base58_str
-                .parse::<Hash>()
-                .map_err(|x| JsValue::from(x.to_string()))
-        } else if let Some(uint8_array) = value.dyn_ref::<Uint8Array>() {
-            <[u8; HASH_BYTES]>::try_from(uint8_array.to_vec())
-                .map(Hash::new_from_array)
-                .map_err(|err| format!("Invalid Hash value: {err:?}").into())
-        } else if let Some(array) = value.dyn_ref::<Array>() {
-            let mut bytes = vec![];
-            let iterator = js_sys::try_iter(&array.values())?.expect("array to be iterable");
-            for x in iterator {
-                let x = x?;
-
-                if let Some(n) = x.as_f64() {
-                    if n >= 0. && n <= 255. {
-                        bytes.push(n as u8);
-                        continue;
-                    }
-                }
-                return Err(format!("Invalid array argument: {:?}", x).into());
-            }
-            <[u8; HASH_BYTES]>::try_from(bytes)
-                .map(Hash::new_from_array)
-                .map_err(|err| format!("Invalid Hash value: {err:?}").into())
-        } else if value.is_undefined() {
-            Ok(Hash::default())
-        } else {
-            Err("Unsupported argument".into())
-        }
-    }
-
-    /// Return the base58 string representation of the hash
-    pub fn toString(&self) -> String {
-        self.to_string()
-    }
-
-    /// Checks if two `Hash`s are equal
-    pub fn equals(&self, other: &Hash) -> bool {
-        self == other
-    }
-
-    /// Return the `Uint8Array` representation of the hash
-    pub fn toBytes(&self) -> Box<[u8]> {
-        self.0.clone().into()
     }
 }
