+++ instruction/src/account_meta.rs
@@ -17,7 +17,6 @@ use solana_pubkey::Pubkey;
 ///
 /// [`Instruction`]: crate::Instruction
 #[repr(C)]
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen::prelude::wasm_bindgen)]
 #[cfg_attr(
     feature = "serde",
     derive(serde_derive::Serialize, serde_derive::Deserialize)
