+++ sdk-wasm-js/src/instruction.rs
@@ -1,0 +1,53 @@
+//! The `Instructions` struct is a legacy workaround
+//! from when wasm-bindgen lacked Vec<T> support
+//! (ref: https://github.com/rustwasm/wasm-bindgen/issues/111)
+#![allow(non_snake_case)]
+
+use {crate::pubkey::Pubkey, wasm_bindgen::prelude::*};
+
+/// wasm-bindgen version of the Instruction struct.
+/// This duplication is required until https://github.com/rustwasm/wasm-bindgen/issues/3671
+/// is fixed. This must not diverge from the regular non-wasm Instruction struct.
+#[wasm_bindgen]
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct Instruction(pub(crate) solana_instruction::Instruction);
+
+#[wasm_bindgen]
+impl Instruction {
+    /// Create a new `Instruction`
+    #[wasm_bindgen(constructor)]
+    pub fn constructor(program_id: Pubkey) -> Self {
+        Instruction(solana_instruction::Instruction::new_with_bytes(
+            program_id.0,
+            &[],
+            std::vec::Vec::new(),
+        ))
+    }
+
+    pub fn setData(&mut self, data: &[u8]) {
+        self.0.data = data.to_vec();
+    }
+
+    pub fn addAccount(&mut self, account_meta: AccountMeta) {
+        self.0.accounts.push(account_meta.0);
+    }
+}
+
+#[wasm_bindgen]
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct AccountMeta(pub(crate) solana_instruction::AccountMeta);
+
+#[wasm_bindgen]
+impl AccountMeta {
+    /// Create a new writable `AccountMeta`
+    pub fn newWritable(pubkey: Pubkey, is_signer: bool) -> Self {
+        AccountMeta(solana_instruction::AccountMeta::new(pubkey.0, is_signer))
+    }
+
+    /// Create a new readonly `AccountMeta`
+    pub fn newReadonly(pubkey: Pubkey, is_signer: bool) -> Self {
+        AccountMeta(solana_instruction::AccountMeta::new_readonly(
+            pubkey.0, is_signer,
+        ))
+    }
+}
