+++ sdk-wasm-js/src/message.rs
@@ -1,0 +1,5 @@
+use wasm_bindgen::prelude::wasm_bindgen;
+
+#[wasm_bindgen]
+#[derive(Default, Debug, PartialEq, Eq, Clone)]
+pub struct Message(pub(crate) solana_message::Message);
