+++ sdk/Cargo.toml
@@ -12,7 +12,7 @@ edition = { workspace = true }
 include = ["src/**/*", "README.md"]
 
 [package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
+targets = ["x86_64-unknown-linux-gnu"]
 all-features = true
 rustdoc-args = ["--cfg=docsrs"]
 
@@ -104,9 +104,6 @@ solana-transaction-error = { workspace = true, features = [
     "serde",
 ], optional = true }
 thiserror = { workspace = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom = { version = "0.1.1", features = ["wasm-bindgen"] }
 
 [dev-dependencies]
 solana-instructions-sysvar = { workspace = true, features = ["dev-context-only-utils"] }
