+++ transaction/Cargo.toml
@@ -10,7 +10,7 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
+targets = ["x86_64-unknown-linux-gnu"]
 all-features = true
 rustdoc-args = ["--cfg=docsrs"]
 
@@ -54,10 +54,6 @@ solana-short-vec = { workspace = true, optional = true }
 solana-signature = { workspace = true }
 solana-signer = { workspace = true, optional = true }
 solana-transaction-error = { workspace = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-solana-keypair = { workspace = true }
-wasm-bindgen = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
